refactor: refactored results page (isolated concerns, reused classes and functions were possible)
